Refactor new aggregation code to reduce the times to call tryConvert.
