+++ sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/plans/logical/basicOperators.scala
@@ -18,6 +18,7 @@
 package org.apache.spark.sql.catalyst.plans.logical
 
 import org.apache.spark.sql.catalyst.expressions._
+import org.apache.spark.sql.catalyst.expressions.aggregate.Utils
 import org.apache.spark.sql.catalyst.plans._
 import org.apache.spark.sql.types._
 import org.apache.spark.util.collection.OpenHashSet
@@ -219,6 +220,8 @@ case class Aggregate(
 
     !expressions.exists(!_.resolved) && childrenResolved && !hasWindowExpressions
   }
+
+  lazy val newAggregation: Option[Aggregate] = Utils.tryConvert(this)
 
   override def output: Seq[Attribute] = aggregateExpressions.map(_.toAttribute)
 }
